requirements.txt
requests
supabase
python-dotenv
pandas
psycopg2-binary
dnspython